T1 1.2.4

+++ T1/src/AFE-API-Domoticz.h
@@ -24,7 +24,7 @@ private:
   char serverURL[184];
   boolean initialized = false;
 
-  const String getApiCall(const char *param, unsigned long idx);
+  const String getApiCall(const char *param, unsigned int idx);
   void callURL(const String url);
 
 public:
@@ -37,12 +37,12 @@ public:
   /* It send to Domoticz switch state using following API call
     json.htm?type=command&param=switchlight&idx=IDX&switchcmd=STATE
   */
-  void sendSwitchCommand(unsigned long idx, const char *value);
+  void sendSwitchCommand(unsigned int idx, const char *value);
 
   /* It send to Domoticz temperature using following API call
      json.htm?type=command&param=udevice&idx=IDX&nvalue=0&svalue=TEMP
   */
-  void sendTemperatureCommand(unsigned long idx, float value);
+  void sendTemperatureCommand(unsigned int idx, float value);
 };
 
 #endif
